Some reverts and pre-PR changes
